Move eval functions around
