+++ nemo_skills/dataset/aai/__init__.py
@@ -12,15 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# settings that define how evaluation should be done by default (all can be changed from cmdline)
-PROMPT_CONFIG = 'generic/math'
-DATASET_GROUP = 'math'
-METRICS_TYPE = "math"
-EVAL_ARGS = "++eval_type=math"
-GENERATION_ARGS = ""
-
-# some answers are not possible to compare symbolically, so have to use a judge model
-# setting openai judge by default, but can be overriden from command line for a locally hosted model
-JUDGE_ARGS = (
-    "--generation_type=math_judge --model=gpt-4.1 --server_type=openai --server_address=https://api.openai.com/v1 "
-)
+IS_BENCHMARK_GROUP = True
