+++ nemo_skills/dataset/hle/__init__.py
@@ -19,3 +19,10 @@ METRICS_TYPE = "math"
 EVAL_ARGS = "++eval_type=math"
 GENERATION_ARGS = ""
 EVAL_SPLIT = "text"
+
+# some answers are not possible to compare symbolically, so have to use a judge model
+# setting openai judge by default, but can be overriden from command line for a locally hosted model
+JUDGE_ARGS = (
+    "++prompt_config=judge/hle ++generation_key=judgement "
+    "--model=gpt-4.1 --server_type=openai --server_address=https://api.openai.com/v1"
+)
